Avoiding out-of-range agreements in S5s
This is not a change in the strategy. It avoids failing due to a bug in
SCML-Std implementation which allowed negotiation issues to exceed
simulation time near the end of the simulation. This bug was fixed in
the version of SCML used for the competition but this fix is necessary
to make S5s work with earlier versions without throwing an exception.